Moving routing_serializer and restriction_loader to routing submodule.
