+++ generator/generator_tests/restriction_collector_test.cpp
@@ -3,7 +3,7 @@
 #include "generator/osm_id.hpp"
 #include "generator/restriction_collector.hpp"
 
-#include "indexer/routing.hpp"
+#include "routing/routing_serializer.hpp"
 
 #include "platform/platform_tests_support/scoped_dir.hpp"
 #include "platform/platform_tests_support/scoped_file.hpp"
