+++ routing/routing_serializer.cpp
@@ -1,4 +1,4 @@
-#include "indexer/routing.hpp"
+#include "routing/routing_serializer.hpp"
 
 namespace
 {
